refined readme.
