+++ comps/finetuning/finetune_runner.py
@@ -2,16 +2,9 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import argparse
-import time
-import uuid
-from typing import List
 
 from llm_on_ray.finetune.finetune_config import FinetuneConfig
 from pydantic_yaml import parse_yaml_raw_as
-from ray.train.base_trainer import TrainingFailedError
-from ray.tune.callback import Callback
-from ray.tune.experiment import Trial
-from ray.tune.logger import LoggerCallback
 from transformers import TrainerCallback, TrainerControl, TrainerState, TrainingArguments
 
 
@@ -38,10 +31,6 @@ def main():
     from llm_on_ray.finetune.finetune import main as llm_on_ray_finetune_main
 
     llm_on_ray_finetune_main(finetune_config)
-    # try:
-    #     llm_on_ray_finetune_main(finetune_config)
-    # except TrainingFailedError as e:
-    #     print(e)
 
 
 if __name__ == "__main__":
